misc cleanup
* split up TCP files into stream / listener
* organize import order
* split up misc other files
