+++ src/net/mod.rs
@@ -7,22 +7,11 @@
 //!
 //! [portability guidelines]: ../struct.Poll.html#portability
 
-mod tcp;
+mod tcp_listener;
+pub use self::tcp_listener::TcpListener;
+
+mod tcp_stream;
+pub use self::tcp_stream::TcpStream;
+
 mod udp;
-
-pub use self::tcp::{TcpListener, TcpStream};
 pub use self::udp::UdpSocket;
-
-#[test]
-#[cfg(not(debug_assertions))]
-fn assert_size() {
-    use std::mem::size_of;
-
-    use crate::sys;
-
-    // Without debug assertions enabled `TcpListener`, `TcpStream` and `UdpSocket` should have the
-    // same size as the system specific socket, i.e. just a file descriptor on Unix platforms.
-    assert_eq!(size_of::<TcpListener>(), size_of::<sys::TcpListener>());
-    assert_eq!(size_of::<TcpStream>(), size_of::<sys::TcpStream>());
-    assert_eq!(size_of::<UdpSocket>(), size_of::<sys::UdpSocket>());
-}
